+++ pkg/storegateway/indexheader/encoding/encoding_test.go
@@ -9,6 +9,8 @@ import (
 	"strconv"
 	"testing"
 
+	"github.com/go-kit/log"
+	"github.com/prometheus/client_golang/prometheus"
 	promencoding "github.com/prometheus/prometheus/tsdb/encoding"
 	"github.com/stretchr/testify/require"
 
@@ -249,7 +251,7 @@ func TestDecbuf_SkipHappyPath(t *testing.T) {
 }
 
 func TestDecbuf_SkipMultipleBufferReads(t *testing.T) {
-	// The underlying FileReader buffers the file 4k bytes at a time. Ensure
+	// The underlying fileReader buffers the file 4k bytes at a time. Ensure
 	// that we can skip multiple 4k chunks without ending up with a short read.
 	bytes := make([]byte, 4096*5)
 	for i := 0; i < len(bytes); i++ {
@@ -432,8 +434,8 @@ func TestDecbuf_UvarintBytesSkipDoesNotCauseBufferFill(t *testing.T) {
 		expectedBytes  = 983
 	)
 
-	// This test verifies that when bytes are read in UvarintBytes, the Peek(n) and
-	// subsequent Skip(len(b)) does not cause a read from disk that invalidates the slice
+	// This test verifies that when bytes are read in UvarintBytes, the peek(n) and
+	// subsequent skip(len(b)) does not cause a read from disk that invalidates the slice
 	// returned. It does this by creating multiple uvarint byte slices in the encoding
 	// buffer each with different content _and_ by ensuring there are more bytes written
 	// in total than the size of the underlying buffer (currently 4k).
@@ -626,10 +628,11 @@ func createDecbufWithBytes(t testing.TB, b []byte) Decbuf {
 	filePath := path.Join(dir, "test-file")
 	require.NoError(t, os.WriteFile(filePath, b, 0700))
 
-	factory := NewDecbufFactory(filePath)
+	reg := prometheus.NewPedanticRegistry()
+	factory := NewDecbufFactory(filePath, 0, log.NewNopLogger(), NewDecbufFactoryMetrics(reg))
 	decbuf := factory.NewRawDecbuf()
 	t.Cleanup(func() {
-		require.NoError(t, factory.Close(decbuf))
+		require.NoError(t, decbuf.Close())
 	})
 
 	require.NoError(t, decbuf.Err())